documentation and rename of file
